src/deployment/deployment_requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mlflow-skinny==2.8.1
xgboost==2.0.2
pandas==2.1.3
scikit-learn==1.3.2
python-dotenv==1.0.0
boto3==1.34.0
prometheus-fastapi-instrumentator==6.1.0
